event_consumer/pyproject.toml
[project]
name = "event-consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "commonmark>=0.9.1",
    "kafka-python>=2.2.15",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.0.0",
    "tenacity>=9.1.2",
]
